+++ R/das_process.R
@@ -101,11 +101,13 @@
 #'     Speed (ship speed, knots)     \tab SpdKt     \tab Event: N; Column: Data2\cr
 #'     Beaufort sea state            \tab Bft       \tab Event: V; Column: Data1\cr
 #'     Swell height (ft)             \tab SwellHght \tab Event: V; Column: Data2\cr
+#'     Swell direction (degrees)     \tab SwellDir  \tab Event: V; Column: Data3\cr
 #'     Wind speed (knots)            \tab WindSpdKt \tab Event: V; Column: Data5\cr
 #'     Rain/fog/haze code            \tab RainFog   \tab Event: W; Column: Data1\cr
 #'     Horizontal sun (clock system) \tab HorizSun  \tab Event: W; Column: Data2\cr
 #'     Vertical sun (clock system)   \tab VertSun   \tab Event: W; Column: Data3\cr
 #'     Glare                         \tab Glare     \tab HorizSun and VertSun\cr
+#'     Wind direction (degrees)      \tab WindDir   \tab Event: W; Column: Data4\cr
 #'     Visibility (nm)               \tab Vis       \tab Event: W; Column: Data5\cr
 #'     Left observer                 \tab ObsL      \tab Event: P; Column: Data1\cr
 #'     Data recorder                 \tab Rec       \tab Event: P; Column: Data2\cr
@@ -303,16 +305,18 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
   SpdKt     <- .process_num(init.val, x, "Data2", event.N, event.na)
   Bft       <- .process_num(init.val, x, "Data1", event.V, event.na)
   SwellHght <- .process_num(init.val, x, "Data2", event.V, event.na)
+  SwellDir  <- .process_num(init.val, x, "Data3", event.V, event.na)
   WindSpdKt <- .process_num(init.val, x, "Data5", event.V, event.na)
   RainFog   <- .process_num(init.val, x, "Data1", event.W, event.na)
   HorizSun  <- .process_num(init.val, x, "Data2", event.W, event.na)
   VertSun   <- .process_num(init.val, x, "Data3", event.W, event.na)
+  WindDir   <- .process_num(init.val, x, "Data4", event.W, event.na)
   Vis       <- .process_num(init.val, x, "Data5", event.W, event.na)
 
-  ObsL <- .process_chr(init.val, x, "Data1", event.P, event.na)
-  Rec  <- .process_chr(init.val, x, "Data2", event.P, event.na)
-  ObsR <- .process_chr(init.val, x, "Data3", event.P, event.na)
-  ObsInd <-  .process_chr(init.val, x, "Data4", event.P, event.na)
+  ObsL   <- .process_chr(init.val, x, "Data1", event.P, event.na)
+  Rec    <- .process_chr(init.val, x, "Data2", event.P, event.na)
+  ObsR   <- .process_chr(init.val, x, "Data3", event.P, event.na)
+  ObsInd <- .process_chr(init.val, x, "Data4", event.P, event.na)
 
   Eff <- as.logical(init.val)
   Eff[sort(unique(c(idx.new.cruise, idx.new.day)))] <- FALSE
@@ -333,8 +337,8 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
     if (i %in% idx.new.cruise) {
       LastEff <- LastEMode <- LastEType <- LastESW <-
         LastCourse <- LastSpdKt <-
-        LastBft <- LastSwH <- LastWSpdKt <-
-        LastRF <- LastHS <- LastVS <- LastVis <-
+        LastBft <- LastSwH <- LastSwD <- LastWSpdKt <-
+        LastRF <- LastHS <- LastVS <- LastWDir <- LastVis <-
         LastObsL <- LastRec <- LastObsR <- LastObsInd <-
         LastCruise <- NA
     }
@@ -343,16 +347,16 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
     if ((i %in% idx.new.day) & reset.day) {
       LastEff <- LastEMode <- LastEType <- LastESW <-
         LastCourse <- LastSpdKt <-
-        LastBft <- LastSwH <- LastWSpdKt <-
-        LastRF <- LastHS <- LastVS <- LastVis <-
+        LastBft <- LastSwH <- LastSwD <- LastWSpdKt <-
+        LastRF <- LastHS <- LastVS <- LastWDir <- LastVis <-
         LastObsL <- LastRec <- LastObsR <- LastObsInd <- NA
     }
 
     # Reset applicable info (all RPVNW-related) when starting BR/R event sequence
     if ((i %in% idx.eff) & reset.effort) {
       LastEType <- LastESW <-  LastCourse <- LastSpdKt <-
-        LastBft <- LastSwH <- LastWSpdKt <-
-        LastRF <- LastHS <- LastVS <- LastVis <-
+        LastBft <- LastSwH <- LastSwD <- LastWSpdKt <-
+        LastRF <- LastHS <- LastVS <- LastWDir <- LastVis <-
         LastObsL <- LastRec <- LastObsR <- LastObsInd <- NA
     }
 
@@ -365,10 +369,12 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
     if (is.na(ESWsides[i]))  ESWsides[i] <- LastESW  else LastESW <- ESWsides[i]  #Sides being surveyed
     if (is.na(Bft[i]))       Bft[i] <- LastBft       else LastBft <- Bft[i]       #Beaufort
     if (is.na(SwellHght[i])) SwellHght[i] <- LastSwH else LastSwH <- SwellHght[i] #Swell height
+    if (is.na(SwellDir[i]))  SwellDir[i] <- LastSwD  else LastSwD <- SwellDir[i] # Swell direction
     if (is.na(WindSpdKt[i])) WindSpdKt[i] <- LastWSpdKt else LastWSpdKt <- WindSpdKt[i] #Wind speed
     if (is.na(RainFog[i]))   RainFog[i] <- LastRF    else LastRF <- RainFog[i]    #Rain or fog
     if (is.na(HorizSun[i]))  HorizSun[i] <- LastHS   else LastHS <- HorizSun[i]   #Horizontal sun
     if (is.na(VertSun[i]))   VertSun[i] <- LastVS    else LastVS <- VertSun[i]    #Vertical sun
+    if (is.na(WindDir[i]))   WindDir[i] <- LastWDir  else LastWDir <- WindDir[i]  #Wind direction
     if (is.na(Vis[i]))       Vis[i] <- LastVis       else LastVis <- Vis[i]       #Visibility
     if (is.na(ObsL[i]))      ObsL[i] <- LastObsL     else LastObsL <- ObsL[i]     #Left obs
     if (is.na(Rec[i]))       Rec[i] <- LastRec       else LastRec <- Rec[i]       #Recorder
@@ -383,7 +389,8 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
   tmp <- list(
     Cruise = Cruise, Mode = Mode,
     EffType = EffType, ESWsides = ESWsides, Course = Course, SpdKt = SpdKt,
-    Bft = Bft, SwellHght = SwellHght, WindSpdKt = WindSpdKt,
+    Bft = Bft, SwellHght = SwellHght, SwellDir = SwellDir,
+    WindSpdKt = WindSpdKt, WindDir = WindDir,
     RainFog = RainFog, HorizSun = HorizSun, VertSun = VertSun, Vis = Vis,
     OnEffort = Eff,
     ObsL = ObsL, Rec = Rec, ObsR = ObsR, ObsInd = ObsInd
@@ -445,7 +452,7 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
   cols.tokeep <- c(
     "Event", "DateTime", "Lat", "Lon", "OnEffort",
     "Cruise", "Mode", "OffsetGMT", "EffType", "ESWsides", "Course", "SpdKt",
-    "Bft", "SwellHght", "WindSpdKt",
+    "Bft", "SwellHght", "SwellDir", "WindSpdKt", "WindDir",
     "RainFog", "HorizSun", "VertSun", "Glare", "Vis",
     "ObsL", "Rec", "ObsR", "ObsInd",
     paste0("Data", 1:12), "EffortDot", "EventNum", "file_das", "line_num"